% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotAvgTemperaturesByYear}
\alias{plotAvgTemperaturesByYear}
\title{Plot of average temperatures by year}
\usage{
plotAvgTemperaturesByYear(avg_temperatures)
}
\description{
\code{plotAvgTemperaturesByYear} plots the average temperature by year in a line graph
}
